+++ test/PerformanceTest.cpp
@@ -1,9 +1,9 @@
 #include "MultiQueueProcessor.h"
 #include "TestStructures.hpp"
 #include <chrono>
+#include <iostream>
 #include <thread>
 #include <vector>
-#include <iostream>
 
 int main(int argc, char **argv) {
   if (argc < 3) {
@@ -24,16 +24,16 @@ int main(int argc, char **argv) {
   MQProcessor::Queue<int, int> mqproc;
   std::vector<TestConsumer> consumers(countCons, TestConsumer());
   std::vector<Providers<BufferSize>> providers;
-  for (int i = 0; i < countCons; i++)
+  for (size_t i = 0; i < countCons; i++)
     mqproc.Subscribe(i, &consumers[i]);
-  for (int i = 0; i < countCons; i++)
+  for (size_t i = 0; i < countCons; i++)
     providers.emplace_back(Providers<BufferSize>(mqproc, countThreads, i));
   for (auto &i : providers)
     i.Start();
   for (auto &i : providers)
     i.Join();
   // std::this_thread::sleep_for(std::chrono::seconds(1));
-  for (int i = 0; i < countCons; i++) {
+  for (size_t i = 0; i < countCons; i++) {
     std::cout << providers[i].Sended() << ", " << consumers[i].consumed << '\n';
   }
   return 0;
